backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.20
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.1.2
spacy==3.8.2
textstat==0.7.4
openai==1.57.4
python-dotenv==1.0.1
